requirements.txt
tensorflow=2.13.0
streamlit
Pillow=10.1.0